Move useChannelsSearch and useChannelsSearchWithLoadMore to hooks, change some imports accordingly
