+++ src/hooks/useChannelsSearchWithLoadMore.ts
@@ -1,31 +1,16 @@
 import { useChannelsList } from "@saleor/channels/queries";
-import { Channel_channel } from "@saleor/channels/types/Channel";
-import { FetchMoreProps, Search, SearchProps } from "@saleor/types";
-import { filter } from "fuzzaldrin";
-import { compact } from "lodash-es";
 import chunk from "lodash-es/chunk";
+import compact from "lodash-es/compact";
 import concat from "lodash-es/concat";
-import React, { useEffect, useState } from "react";
+import { useEffect, useState } from "react";
+
+import {
+  ChannelsWithLoadMoreProps,
+  useChannelsSearch
+} from "./useChannelsSearch";
 
 const DEFAULT_ITEMS_PER_PAGE = 6;
 const INITIAL_INDEX = 0;
-
-export const useChannelsSearch = function<T extends { name: string }>(
-  channels: T[]
-) {
-  const [query, onQueryChange] = React.useState("");
-  const filteredChannels =
-    filter<T, "name">(channels, query, { key: "name" }) || [];
-
-  return { query, onQueryChange, filteredChannels };
-};
-
-export interface ChannelsWithLoadMoreProps
-  extends FetchMoreProps,
-    Search,
-    SearchProps {
-  channels: Channel_channel[];
-}
 
 export const useChannelsSearchWithLoadMore = (
   itemsPerPage: number = DEFAULT_ITEMS_PER_PAGE
